urfa_link/requirements.txt
fastapi
uvicorn
pydantic
numpy==1.26.4
google-generativeai==0.4.1
cryptography==42.0.5
haversine==2.8.1
sqlalchemy>=2.0.30
httpx>=0.27.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
python-multipart
